+++ transport/http/handler.go
@@ -47,9 +47,21 @@ type handler struct {
 	router      transport.Router
 	tracer      opentracing.Tracer
 	grabHeaders map[string]struct{}
+	override    *override
+}
+
+type override struct {
+	check   func(req *http.Request) bool
+	handler http.Handler
 }
 
 func (h handler) ServeHTTP(w http.ResponseWriter, req *http.Request) {
+	// if override is provided and check returns true, execute override handler
+	if h.override != nil && h.override.check(req) {
+		h.override.handler.ServeHTTP(w, req)
+		return
+	}
+
 	responseWriter := newResponseWriter(w)
 	service := popHeader(req.Header, ServiceHeader)
 	procedure := popHeader(req.Header, ProcedureHeader)